mongodb installed and config
