+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/scailancrei/FullStackOpen-3.0#readme",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "mongoose": "^8.5.1",
     "morgan": "^1.10.0"
